+++ source/Delphi.Forms.Types.pas
@@ -1,0 +1,144 @@
+unit Delphi.Forms.Types;
+
+interface
+
+uses
+  System.SysUtils,
+  System.Generics.Collections;
+
+type
+
+  TObjectKind = (okObject, okInherited, okInline);
+
+  TFormValueKind = (
+    fvInteger,
+    fvFloat,
+    fvString,
+    fvBoolean,
+    fvIdentifier,
+    fvSet,
+    fvBinary,
+    fvList,
+    fvCollection
+  );
+
+  TFormValue = class;
+  TFormProperty = class;
+  TFormObject = class;
+
+  TFormValueList = TObjectList<TFormValue>;
+  TFormPropertyList = TObjectList<TFormProperty>;
+  TFormObjectList = TObjectList<TFormObject>;
+
+  TFormValue = class
+  public
+    Kind: TFormValueKind;
+    IntValue: Int64;
+    FloatValue: Extended;
+    StringValue: string;
+    BoolValue: Boolean;
+    IdentValue: string;
+    SetItems: TArray<string>;
+    BinaryData: TBytes;
+    BinaryHex: string;
+    ListItems: TFormValueList;
+    CollectionItems: TFormObjectList;
+    RawText: string;
+    constructor Create(AKind: TFormValueKind);
+    destructor Destroy; override;
+  end;
+
+  TFormProperty = class
+  public
+    Name: string;
+    Value: TFormValue;
+    constructor Create(const AName: string; AValue: TFormValue);
+    destructor Destroy; override;
+  end;
+
+  TFormObject = class
+  public
+    ObjectKind: TObjectKind;
+    Name: string;
+    ClassName_: string;
+    Properties: TFormPropertyList;
+    Children: TFormObjectList;
+    constructor Create;
+    destructor Destroy; override;
+  end;
+
+  TFormFile = class
+  public
+    Root: TFormObject;
+    constructor Create;
+    destructor Destroy; override;
+  end;
+
+implementation
+
+{ TFormValue }
+
+constructor TFormValue.Create(AKind: TFormValueKind);
+begin
+  inherited Create;
+  Kind := AKind;
+  case AKind of
+    fvList:
+      ListItems := TFormValueList.Create;
+    fvCollection:
+      CollectionItems := TFormObjectList.Create;
+  end;
+end;
+
+destructor TFormValue.Destroy;
+begin
+  ListItems.Free;
+  CollectionItems.Free;
+  inherited;
+end;
+
+{ TFormProperty }
+
+constructor TFormProperty.Create(const AName: string; AValue: TFormValue);
+begin
+  inherited Create;
+  Name := AName;
+  Value := AValue;
+end;
+
+destructor TFormProperty.Destroy;
+begin
+  Value.Free;
+  inherited;
+end;
+
+{ TFormObject }
+
+constructor TFormObject.Create;
+begin
+  inherited Create;
+  Properties := TFormPropertyList.Create;
+  Children := TFormObjectList.Create;
+end;
+
+destructor TFormObject.Destroy;
+begin
+  Children.Free;
+  Properties.Free;
+  inherited;
+end;
+
+{ TFormFile }
+
+constructor TFormFile.Create;
+begin
+  inherited Create;
+end;
+
+destructor TFormFile.Destroy;
+begin
+  Root.Free;
+  inherited;
+end;
+
+end.